添加维度限制，电压限制
